Fix some type issues building on x64 Linux.

+++ src/sh/shuos.c
@@ -235,7 +235,7 @@ Return Value:
 
 {
 
-    *Handle = (void *)FileDescriptorToRead;
+    *Handle = (void *)(long)FileDescriptorToRead;
     return 1;
 }
 
@@ -283,7 +283,7 @@ Return Value:
     Buffer = NULL;
     BufferCapacity = 0;
     BufferSize = 0;
-    Descriptor = (int)Handle;
+    Descriptor = (int)(long)Handle;
     Result = 1;
 
     //
